Fix the sigev struct to support musl libc
Fixes #19

It appears musl libc has actually added SIGEV_THREAD_ID support. However, the
member name for the thread id in the sigev struct is different than
glibc's. Musl provides a macro with this name (sigev_notify_thread_id), glibc
does not.

Support both by defining this macro if it does not exist. If it does not exist,
the value of the macro is the name of the member in glibc. Then, use this macro
in the subsequent struct definition.

https://git.musl-libc.org/cgit/musl/commit/?id=7c71792e87691451f2a6b76348e83ad1889f1dcb