osicat.asd: use feature expressions rather than reader conditionals
The ASDF Manual recommends this because it means that code can extract
information about what components and dependencies the system would have on
platforms other than the one the code is running on.

Another possibility this enables is temporarily binding *FEATURES* to match a
target system which is different from the current system, and then using an
operation like ASDF:MONOLITHIC-CONCATENATE-SOURCE-OP to prepare a number of
dependencies for loading up into Lisp on the target system.