Split some parts of osicat.lisp into basic-osicat.lisp and windows.lisp
Some functionality of osicat.lisp cannot be compiled on Windows without a
WARNING being signaled, due to missing POSIX functions. Move everything that
works on Windows into basic-osicat.lisp, keep everything else in
osicat.lisp (which is not loaded on Windows) and add a windows.lisp file that
marks functions as unsupported.