windows: Improve temporary file support
First, look at TMP instead of TMPDIR on Windows.

Second, it's not possible to unlink a file if someone has it open, unless you
use CreateFileA/W and set FILE_SHARE_DELETE. Therefore, note that
OPEN-TEMPORARY-FILE may return a pathname as a second value. If it does, the
user is reponsible to delete that file.

Third, modify WITH-TEMPORARY-FILE to unlink the file.