make-link: Implement on Windows and fix on non-Windows
On non-Windows platforms there is no need to change the current directory; we
can just merge with *D-P-D* and absolutize.

Additionally, specify that the target is merged with the link's
directory. Otherwise, it's possible to inherit the name or type from link if it
is unspecified in target, which is probably not what anyone wants.